test(LineAcceptanceTest): 구간 관련 테스트 추가
