+++ src/test/java/wooteco/subway/line/LineAcceptanceTest.java
@@ -6,6 +6,7 @@ import static wooteco.subway.station.StationAcceptanceTest.지하철역_등록;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
+import java.net.URI;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -15,20 +16,22 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import wooteco.subway.AcceptanceTest;
+import wooteco.subway.line.section.SectionRequest;
+import wooteco.subway.line.section.SectionResponse;
 import wooteco.subway.station.StationRequest;
 
 @DisplayName("노선 관련 기능")
 public class LineAcceptanceTest extends AcceptanceTest {
 
-    public static final LineRequest LINE_2 = new LineRequest(
+    private final LineRequest LINE_2 = new LineRequest(
         "2호선", "grey darken-1", 1L, 2L, 2, 500
     );
 
-    public static final LineRequest LINE_3 = new LineRequest(
+    private final LineRequest LINE_3 = new LineRequest(
         "3호선", "grey darken-2", 5L, 6L, 12, 1500
     );
 
-    public static ExtractableResponse<Response> 노선_등록(final LineRequest lineRequest) {
+    private ExtractableResponse<Response> 노선_등록(final LineRequest lineRequest) {
         return RestAssured.given().log().all()
             .body(lineRequest)
             .contentType(MediaType.APPLICATION_JSON_VALUE)
@@ -207,5 +210,60 @@ public class LineAcceptanceTest extends AcceptanceTest {
         final ExtractableResponse<Response> response = 노선_제거("/lines/2000000");
 
         assertThat(response.statusCode()).isEqualTo(HttpStatus.NOT_FOUND.value());
+    }
+
+    @DisplayName("구간을 생성한다.")
+    @Test
+    void createSection() {
+        final String uri = 노선_등록(LINE_2).header("Location") + "/sections";
+        final SectionRequest sectionRequest = new SectionRequest(2L, 4L, 10);
+        final ExtractableResponse<Response> response = 구간_등록(uri, sectionRequest);
+
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+        assertThat(response.header("Location")).isNotBlank();
+        구간_생성값_검증(response, sectionRequest);
+    }
+
+    private ExtractableResponse<Response> 구간_등록(final String uri, final SectionRequest sectionRequest) {
+        return RestAssured.given().log().all()
+            .body(sectionRequest)
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .when()
+            .post(uri)
+            .then().log().all()
+            .extract();
+    }
+
+    private void 구간_생성값_검증(final ExtractableResponse<Response> response, final SectionRequest sectionRequest) {
+        final SectionResponse sectionResponse = response.body().as(SectionResponse.class);
+        assertThat(sectionResponse.getId()).isEqualTo(getCreatedSectionId(response));
+        assertThat(sectionResponse.getDownStationId()).isEqualTo(sectionRequest.getDownStationId());
+        assertThat(sectionResponse.getUpStationId()).isEqualTo(sectionRequest.getUpStationId());
+        assertThat(sectionResponse.getDistance()).isEqualTo(sectionRequest.getDistance());
+    }
+
+    private long getCreatedSectionId(final ExtractableResponse<Response> response) {
+        return Long.parseLong(response.header("Location").split("/")[4]);
+    }
+
+    @DisplayName("구간을 제거한다.")
+    @Test
+    void deleteSection() {
+        final String uri = 노선_등록(LINE_2).header("Location") + "/sections";
+        final SectionRequest leftSectionRequest = new SectionRequest(2L, 4L, 10);
+        final SectionRequest rightSectionRequest = new SectionRequest(4L, 6L, 10);
+        구간_등록(uri, leftSectionRequest);
+        구간_등록(uri, rightSectionRequest);
+
+        final ExtractableResponse<Response> response = 구간_제거(String.format("%s?stationId=%s", uri, 4L));
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
+    }
+
+    private ExtractableResponse<Response> 구간_제거(final String uri) {
+        return RestAssured.given().log().all()
+            .when()
+            .delete(uri)
+            .then().log().all()
+            .extract();
     }
 }
